pyproject.toml
[tool.poetry]
name = "goryaev-ceramis"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "goryaev_ceramis"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.95.1"}
alembic = "^1.10.4"
psycopg2 = "^2.9.6"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.15"}
python-dotenv = "^1.0.0"
pydantic = "^1.10.8"
asyncpg = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aiofiles = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
